# api_docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "More-Recipes provides a platform for user to share the awesome and exciting  recipe ideas they have invented or learnt.  Suppose a user comes up with a recipe,  he/she can post it on More-Recipes and  get feedback in form of reviews and votes from other user who explore that recipe. user can also keep a list of their favorite recipes on the application.",
    "version": "v1",
    "title": "More Recipes API",
    "contact": {
      "email": "ebukeokonji@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/swagger-api/swagger-ui/blob/master/LICENSE"
    }
  },
  "basePath": "/api/v1/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/users/reset-password": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user reset password",
        "description": "A user can reset password",
        "operationId": "resetPassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successful"
          },
          "403": {
            "description": "You`re unauthorized to perform this action"
          },
          "404": {
            "description": "User not found"
          },
          "default": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/change-password": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Change User Password",
        "description": "A user can change password",
        "operationId": "changePassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "description": "Authentication Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Password changed"
          },
          "400": {
            "description": "Invalid password"
          },
          "default": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}